+++ test/utils/e2e_helpers.go
@@ -83,7 +83,7 @@ func ForceDeleteNamespace(namespace string) error {
 	resourceTypes := []string{
 		"homeassistants", "homeassistantconfigurations",
 		"homeassistantsecrets", "homeassistantautomations",
-		"homeassistantscenes", "homeassistantscripts", "homeassistantaddons",
+		"homeassistantscenes", "homeassistantscripts",
 	}
 
 	for _, resourceType := range resourceTypes {